+++ packages/tokens/src/1101.ts
@@ -15,4 +15,13 @@ export const polygonZkEvmTokens = {
     'Matic Token',
     'https://polygon.technology/polygon-zkevm',
   ),
+  grai: new ERC20Token(
+    ChainId.POLYGON_ZKEVM,
+    '0xCA68ad4EE5c96871EC6C6dac2F714a8437A3Fe66',
+    18,
+    'GRAI',
+    'Gravita Debt Token',
+    'https://www.gravitaprotocol.com/',
+  ),
+  wbtc: new ERC20Token(ChainId.POLYGON_ZKEVM, '0xEA034fb02eB1808C2cc3adbC15f447B93CbE08e1', 8, 'WBTC', 'Wrapped BTC'),
 }